function that executes a function

+++ function_pointers/1-array_iterator.c
@@ -9,8 +9,8 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
-	
+	size_t i;
+
 	for (i = 0; i < size; i++)
 	{
 		action(array[i]);
